requirements.txt
absl-py==0.10.0
addict==2.3.0
backports.functools-lru-cache==1.6.1
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20

chardet==3.0.4
click==7.1.2

cycler==0.10.0
Cython==0.29.21
cytoolz==0.11.0

decorator==4.4.2
efficientnet-pytorch==0.7.0
faiss-cpu==1.6.3
filelock==3.0.12
future==0.18.2
gdown==3.12.2
google-auth==1.21.2
google-auth-oauthlib==0.4.1

h5py==2.10.0
idna==2.10

imgaug==0.4.0
importlib-metadata==1.7.0
ipython==5.8.0
ipython-genutils==0.2.0
joblib==0.16.0

Markdown==3.2.2

mmcv==1.1.5

numpy==1.18.5
oauthlib==3.0.1
olefile==0.46
opencv-python==4.4.0.44
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==1.0.15
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8

PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-ignite==0.1.2
pandas==1.2.3
PyYAML==5.3.1
requests==2.24.0

rsa==4.6
scikit-image==0.16.2
scikit-learn==0.23.2
scipy==1.9.0

simplegeneric==0.8.1
six==1.15.0
sklearn==0.0
tabulate==0.8.7

tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
threadpoolctl==2.1.0
timm==0.3.4

# provisioning by the container itself
#torch==1.6.0
#torchvision==0.7.0

tqdm==4.49.0
traitlets==4.3.3
urllib3==1.25.10
visdom==0.1.8.9
Werkzeug==1.0.1
yacs==0.1.8
yapf==0.30.0
zipp==3.1.0
